report-service

+++ app/services/report_service.py
@@ -200,7 +200,7 @@ async def daily_booking_report(
         "total_amount": total_amount,
         "items": items,
         "totals": {
-            "amount": _to_float(opening_amount + total_amount)
+            "amount": _to_float(float(opening_amount) + float(total_amount))
         }
     }
 
@@ -257,8 +257,8 @@ async def customer_wise_booking_report(
         "opening_pending_amount": opening_vals[1],
         "items": items,
         "totals": {
-            "revenue": _to_float(opening_vals[0] + total_rev),
-            "pending_amount": _to_float(opening_vals[1] + total_pending)
+            "revenue": _to_float(float(opening_vals[0]) + float(total_rev)),
+            "pending_amount": _to_float(float(opening_vals[1]) + float(total_pending))
         }
     }
 
@@ -301,7 +301,7 @@ async def service_type_report(
         "opening_amount": opening_revenue,
         "items": items,
         "totals": {
-            "revenue": _to_float(opening_revenue + total_rev)
+            "revenue": _to_float(float(opening_revenue) + float(total_rev))
         }
     }
 
@@ -431,7 +431,7 @@ async def cod_pending_report(
         "total_pending_amount": total_pending,
         "items": items,
         "totals": {
-            "pending_amount": _to_float(opening_pending + total_pending)
+            "pending_amount": _to_float(float(opening_pending) + float(total_pending))
         }
     }
 
@@ -488,11 +488,11 @@ async def gst_sales_report(
         "opening_total": opening_vals[2],
         "items": items,
         "totals": {
-            "taxable_amount": _to_float(opening_vals[0] + total_taxable),
-            "cgst": _to_float(_to_float(opening_vals[1] / 2) + total_cgst),
-            "sgst": _to_float(_to_float(opening_vals[1] / 2) + total_sgst),
+            "taxable_amount": _to_float(float(opening_vals[0]) + float(total_taxable)),
+            "cgst": _to_float(float(_to_float(opening_vals[1] / 2)) + float(total_cgst)),
+            "sgst": _to_float(float(_to_float(opening_vals[1] / 2)) + float(total_sgst)),
             "igst": total_igst,
-            "total": _to_float(opening_vals[2] + total_total)
+            "total": _to_float(float(opening_vals[2]) + float(total_total))
         }
     }
 
@@ -564,10 +564,10 @@ async def franchise_settlement_report(
         "opening_net_payable": opening_franchise,
         "items": items,
         "totals": {
-            "revenue": _to_float(opening_rev_val + total_revenue),
-            "ho_share": _to_float(opening_ho + total_ho),
-            "franchise_share": _to_float(opening_franchise + total_franchise),
-            "net_payable": _to_float(opening_franchise + total_payable)
+            "revenue": _to_float(float(opening_rev_val) + float(total_revenue)),
+            "ho_share": _to_float(float(opening_ho) + float(total_ho)),
+            "franchise_share": _to_float(float(opening_franchise) + float(total_franchise)),
+            "net_payable": _to_float(float(opening_franchise) + float(total_payable))
         }
     }
 
@@ -612,9 +612,9 @@ async def monthly_revenue_analysis(db: AsyncSession, current_user: User, year: i
         "opening_profit": _to_float(opening_revenue),
         "items": items,
         "totals": {
-            "revenue": _to_float(opening_revenue + total_revenue),
+            "revenue": _to_float(float(opening_revenue) + float(total_revenue)),
             "expenses": total_expenses,
-            "profit": _to_float(opening_revenue + total_profit)
+            "profit": _to_float(float(opening_revenue) + float(total_profit))
         }
     }
 
@@ -634,8 +634,8 @@ async def top_customer_report(
     total_rev = _to_float(sum(item["revenue"] for item in data["items"]))
     total_pending = _to_float(sum(item["pending_amount"] for item in data["items"]))
     data["totals"] = {
-        "revenue": _to_float(data.get("opening_revenue", 0.0) + total_rev),
-        "pending_amount": _to_float(data.get("opening_pending_amount", 0.0) + total_pending)
+        "revenue": _to_float(float(data.get("opening_revenue", 0.0)) + float(total_rev)),
+        "pending_amount": _to_float(float(data.get("opening_pending_amount", 0.0)) + float(total_pending))
     }
     return data
 
@@ -759,7 +759,7 @@ async def day_close_report(
     
     total_collection = _to_float(today_collection)
     total_expenses = _to_float(float(today_payments) + float(today_expenses_table))
-    closing_balance = _to_float(opening_balance + total_collection - total_expenses)
+    closing_balance = _to_float(float(opening_balance) + float(total_collection) - float(total_expenses))
     
     items = [{
         "opening_balance": opening_balance,
@@ -863,7 +863,7 @@ async def branch_activity_report(
         "totals": {
             "bookings": total_bookings,
             "deliveries": total_deliveries,
-            "collections": _to_float(opening_collections + total_collections),
+            "collections": _to_float(float(opening_collections) + float(total_collections)),
             "pending": total_pending
         }
     }
@@ -1007,7 +1007,7 @@ async def collection_summary_report(
         "opening_amount": opening_amount,
         "items": items,
         "totals": {
-            "amount": _to_float(opening_amount + total_amount)
+            "amount": _to_float(float(opening_amount) + float(total_amount))
         }
     }
 
@@ -1055,7 +1055,7 @@ async def outstanding_collection_report(
         "opening_balance": opening_balance,
         "items": items,
         "totals": {
-            "balance": _to_float(opening_balance + total_balance)
+            "balance": _to_float(float(opening_balance) + float(total_balance))
         }
     }
 
@@ -1144,10 +1144,10 @@ async def daily_collection_report(
         "opening_total": _to_float(opening_total),
         "items": items,
         "totals": {
-            "cash": _to_float(opening_cash + total_cash),
-            "upi": _to_float(opening_upi + total_upi),
-            "bank_transfer": _to_float(opening_bank + total_bank),
-            "total": _to_float(opening_total + total_total)
+            "cash": _to_float(float(opening_cash) + float(total_cash)),
+            "upi": _to_float(float(opening_upi) + float(total_upi)),
+            "bank_transfer": _to_float(float(opening_bank) + float(total_bank)),
+            "total": _to_float(float(opening_total) + float(total_total))
         }
     }
 
@@ -1220,9 +1220,9 @@ async def cod_settlement_report(
         "opening_net_payable": opening_net_val,
         "items": items,
         "totals": {
-            "cod_amount": _to_float(opening_cod_val + total_cod),
-            "commission": _to_float(opening_comm_val + total_commission),
-            "net_payable": _to_float(opening_net_val + total_payable)
+            "cod_amount": _to_float(float(opening_cod_val) + float(total_cod)),
+            "commission": _to_float(float(opening_comm_val) + float(total_commission)),
+            "net_payable": _to_float(float(opening_net_val) + float(total_payable))
         }
     }
 
@@ -1282,7 +1282,7 @@ async def cod_commission_report(
         "opening_amount": _to_float(opening_cod * 0.05),
         "items": items,
         "totals": {
-            "amount": _to_float((opening_cod * 0.05) + total_commission)
+            "amount": _to_float(float(opening_cod * 0.05) + float(total_commission))
         }
     }
 
@@ -1406,7 +1406,7 @@ async def expense_report(
         "opening_amount": opening_amount,
         "items": items,
         "totals": {
-            "amount": _to_float(opening_amount + total_amount)
+            "amount": _to_float(float(opening_amount) + float(total_amount))
         }
     }
 
@@ -1635,9 +1635,9 @@ async def gst_collection_summary(
         "opening_balance": opening_balance,
         "items": items,
         "totals": {
-            "collected_gst": _to_float(opening_collected + total_collected),
-            "paid_gst": _to_float(opening_paid + total_paid),
-            "balance": _to_float(opening_balance + total_balance)
+            "collected_gst": _to_float(float(opening_collected) + float(total_collected)),
+            "paid_gst": _to_float(float(opening_paid) + float(total_paid)),
+            "balance": _to_float(float(opening_balance) + float(total_balance))
         }
     }
 
@@ -1685,7 +1685,7 @@ async def franchise_outstanding_report(
         "opening_pending_amount": opening_pending,
         "items": items,
         "totals": {
-            "pending_amount": _to_float(opening_pending + total_pending)
+            "pending_amount": _to_float(float(opening_pending) + float(total_pending))
         }
     }
 
@@ -1802,9 +1802,9 @@ async def franchise_collection_report(
         "opening_balance": opening_balance_val,
         "items": items,
         "totals": {
-            "cash_collection": _to_float(total_cash_opening + total_cash),
-            "bank_deposit": _to_float(total_bank_opening + total_bank),
-            "closing_balance": _to_float(opening_balance_val + total_closing)
+            "cash_collection": _to_float(float(total_cash_opening) + float(total_cash)),
+            "bank_deposit": _to_float(float(total_bank_opening) + float(total_bank)),
+            "closing_balance": _to_float(float(opening_balance_val) + float(total_closing))
         }
     }
 
@@ -1898,9 +1898,9 @@ async def franchise_profitability_report(
         "opening_profit": opening_profit_val,
         "items": items,
         "totals": {
-            "revenue": _to_float(opening_rev_val + total_revenue),
-            "expenses": _to_float(opening_exp_val + total_expenses),
-            "profit": _to_float(opening_profit_val + total_profit)
+            "revenue": _to_float(float(opening_rev_val) + float(total_revenue)),
+            "expenses": _to_float(float(opening_exp_val) + float(total_expenses)),
+            "profit": _to_float(float(opening_profit_val) + float(total_profit))
         }
     }
 
@@ -1975,7 +1975,7 @@ async def area_wise_business_report(
         "items": items,
         "totals": {
             "shipments": total_shipments,
-            "revenue": _to_float(opening_revenue + total_revenue),
+            "revenue": _to_float(float(opening_revenue) + float(total_revenue)),
             "pending_deliveries": total_pending
         }
     }
